Refactor: Modularize wizard initialization & validation
- Extracted wizard initialization logic into useWizardInit hook.
- Moved save functionality to useSaveWizardStep hook.
- Improved separation of concerns and overall maintainability.
